+++ src/Core/File/FilePathGenerator.php
@@ -54,4 +54,5 @@ readonly class FilePathGenerator
 
         return $path;
     }
+
 }