+++ src/Core/JobType.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace AppTank\Horus\Core;
+
+enum JobType
+{
+    case GENERATE_SYNC_DATA;
+}