+++ src/Illuminate/Job/GenerateDataSyncJob.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace AppTank\Horus\Illuminate\Job;
+
+use AppTank\Horus\Core\Auth\UserAuth;
+use AppTank\Horus\Core\Config\Config;
+use AppTank\Horus\Core\File\IFileHandler;
+use AppTank\Horus\Core\Model\SyncJob;
+use AppTank\Horus\Core\Repository\IGetDataEntitiesUseCase;
+use AppTank\Horus\Core\Repository\SyncJobRepository;
+use AppTank\Horus\Core\SyncJobStatus;
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
+
+/**
+ * @internal Class GenerateDataSyncJob
+ *
+ * Handles the generation of data synchronization jobs.
+ * This job retrieves data entities and saves them to a file for download.
+ *
+ * @package AppTank\Horus\Illuminate\Job
+ *
+ * @author John Ospina
+ * Year: 2024
+ */
+class GenerateDataSyncJob implements ShouldQueue
+{
+    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
+
+    function __construct(
+        private readonly UserAuth $userAuth,
+        private readonly SyncJob  $syncJob
+    )
+    {
+    }
+
+    public function handle(
+        IGetDataEntitiesUseCase $getDataEntitiesUseCase,
+        SyncJobRepository       $syncJobRepository,
+        IFileHandler            $fileHandler,
+        Config                  $config
+    ): void
+    {
+        // Update the job status to IN_PROGRESS
+        $jobInProgress = $this->syncJob->cloneWithStatus(SyncJobStatus::IN_PROGRESS);
+        $syncJobRepository->save($jobInProgress);
+
+        try {
+            // Get the data entities using the repository
+            $data = json_encode($getDataEntitiesUseCase->__invoke($this->userAuth, $this->syncJob->checkpoint));
+            $pathFile = $config->getPathFilesSync() . "/{$this->syncJob->id}.json";
+            $fileUrl = $fileHandler->createDownloadableTemporaryFile($pathFile, $data, "application/json");
+
+            // Update the job with the download URL and result timestamp
+            $jobCompleted = new SyncJob(
+                $this->syncJob->id,
+                $this->syncJob->userId,
+                SyncJobStatus::SUCCESS,
+                resultAt: now()->toImmutable(),
+                downloadUrl: $fileUrl,
+                checkpoint: $this->syncJob->checkpoint
+            );
+            $syncJobRepository->save($jobCompleted);
+        } catch (\Throwable $e) {
+            report($e);
+            $syncJobRepository->save($this->syncJob->cloneWithStatus(SyncJobStatus::FAILED));
+        }
+
+    }
+}