+++ src/Illuminate/Provider/HorusServiceProvider.php
@@ -5,26 +5,33 @@ namespace AppTank\Horus\Illuminate\Provider;
 use AppTank\Horus\Client\HorusQueueActionClient;
 use AppTank\Horus\Client\IHorusQueueActionClient;
 use AppTank\Horus\Core\Bus\IEventBus;
+use AppTank\Horus\Core\Bus\IJobDispatcher;
+use AppTank\Horus\Core\File\FilePathGenerator;
 use AppTank\Horus\Core\Mapper\EntityMapper;
 use AppTank\Horus\Core\Repository\CacheRepository;
 use AppTank\Horus\Core\Repository\EntityAccessValidatorRepository;
 use AppTank\Horus\Core\Repository\EntityRepository;
 use AppTank\Horus\Core\Repository\FileUploadedRepository;
+use AppTank\Horus\Core\Repository\IGetDataEntitiesUseCase;
 use AppTank\Horus\Core\Repository\MigrationSchemaRepository;
 use AppTank\Horus\Core\Repository\QueueActionRepository;
+use AppTank\Horus\Core\Repository\SyncJobRepository;
 use AppTank\Horus\Core\Transaction\ITransactionHandler;
 use AppTank\Horus\Core\Util\IDateTimeUtil;
 use AppTank\Horus\Horus;
 use AppTank\Horus\Illuminate\Bus\EventBus;
+use AppTank\Horus\Illuminate\Bus\JobDispatcher;
 use AppTank\Horus\Illuminate\Console\CreateEntitySynchronizableCommand;
 use AppTank\Horus\Illuminate\Console\PruneFilesUploadedCommand;
 use AppTank\Horus\Illuminate\Transaction\EloquentTransactionHandler;
 use AppTank\Horus\Illuminate\Util\DateTimeUtil;
+use AppTank\Horus\Application\Get\GetDataEntities;
 use AppTank\Horus\Repository\DefaultCacheRepository;
 use AppTank\Horus\Repository\EloquentEntityAccessValidatorRepository;
 use AppTank\Horus\Repository\EloquentEntityRepository;
 use AppTank\Horus\Repository\EloquentFileUploadedRepository;
 use AppTank\Horus\Repository\EloquentQueueActionRepository;
+use AppTank\Horus\Repository\EloquentSyncJobRepository;
 use AppTank\Horus\Repository\StaticMigrationSchemaRepository;
 use Carbon\Laravel\ServiceProvider;
 use Illuminate\Support\Facades\Route;
@@ -67,6 +74,10 @@ class HorusServiceProvider extends ServiceProvider
 
         $this->app->singleton(IEventBus::class, function () {
             return new EventBus();
+        });
+
+        $this->app->singleton(IJobDispatcher::class, function () {
+            return new JobDispatcher();
         });
 
         $this->app->bind(EntityMapper::class, function () {
@@ -113,6 +124,10 @@ class HorusServiceProvider extends ServiceProvider
             return new EloquentFileUploadedRepository();
         });
 
+        $this->app->singleton(SyncJobRepository::class, function () {
+            return new EloquentSyncJobRepository();
+        });
+
         $this->app->bind(IHorusQueueActionClient::class, function () {
             return new HorusQueueActionClient(
                 $this->app->make(ITransactionHandler::class),
@@ -122,10 +137,24 @@ class HorusServiceProvider extends ServiceProvider
             );
         });
 
+
+        $this->app->singleton(FilePathGenerator::class, function () {
+            return new FilePathGenerator(
+                $this->app->make(EntityRepository::class),
+                Horus::getInstance()->getConfig()
+            );
+        });
+
         $this->app->singleton(CacheRepository::class, function () {
             return new DefaultCacheRepository();
         });
 
+        $this->app->bind(IGetDataEntitiesUseCase::class, function () {
+            return new GetDataEntities(
+                $this->app->make(EntityRepository::class),
+                $this->app->make(EntityAccessValidatorRepository::class)
+            );
+        });
 
     }